feat: add a confirmation before downloading

+++ main.sh
@@ -32,25 +32,42 @@ EXISTING_JAR=$(ls patrimoine@*.jar 2>/dev/null | head -n 1)
 if [ -n "$EXISTING_JAR" ]; then
     CURRENT_VERSION=$(echo "$EXISTING_JAR" | cut -d "@" -f 2 | cut -d "." -f 1-3)
     if [ "$CURRENT_VERSION" = "$LAST_VERSION" ]; then
-        echo "You already have the latest version : $EXISTING_JAR in $(pwd)/$EXISTING_JAR"
+        echo "You already have the latest version : $EXISTING_JAR"
     else
-        echo "Old version detected : $CURRENT_VERSION. Deleting and downloading the version $LAST_VERSION..."
-        rm -f "$EXISTING_JAR"
+        echo "Old version detected : $CURRENT_VERSION."
+        read -p "Update to version $LAST_VERSION? (y/N) : " -r
+        if [[ $REPLY =~ ^[OoYy]$ ]]; then
+            echo "Updating..."
+            rm -f "$EXISTING_JAR"
+            if is_windows; then
+                curl --ssl-no-revoke -L -o "$JAR_NAME" "$LINK"
+            else
+                curl -L -o "$JAR_NAME" "$LINK"
+            fi
+        else
+            echo "Update skipped. Using existing version $CURRENT_VERSION."
+            JAR_NAME="$EXISTING_JAR"
+        fi
+    fi
+else
+    echo "No version detected."
+    read -p "Download version $LAST_VERSION? (y/N) : " -r
+    if [[ $REPLY =~ ^[OoYy]$ ]]; then
+        echo "Downloading..."
         if is_windows; then
             curl --ssl-no-revoke -L -o "$JAR_NAME" "$LINK"
         else
             curl -L -o "$JAR_NAME" "$LINK"
         fi
-    fi
-else
-    echo "No version detected. Downloading version $LAST_VERSION..."
-    echo "The download may take a few moments..."
-    if is_windows; then
-        curl --ssl-no-revoke -L -o "$JAR_NAME" "$LINK"
     else
-        curl -L -o "$JAR_NAME" "$LINK"
+        echo "Download canceled. Cannot launch Patrimoine."
+        exit 0
     fi
 fi
 
-echo "Launching $JAR_NAME..."
-java -Dpatrimoine.mode=OFFLINE -jar "$JAR_NAME"
+if [ -f "$JAR_NAME" ]; then
+    echo "Launching $JAR_NAME..."
+    java -Dpatrimoine.mode=OFFLINE -jar "$JAR_NAME"
+else
+    echo "ERROR: JAR file not found. Launch aborted."
+fi
